Safari-611.1.10/LayoutTests/http/tests/loading/preload-css-with-csp-nonce-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
This test makes sure that the Preload scanner preloads a stylesheet resource with a nonce attribute.
PASS
